@@ -2,6 +2,7 @@
   "name": "template",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -16,8 +17,10 @@
     "@emotion/styled": "^11",
     "@sendgrid/mail": "^8.0.0",
     "@vercel/analytics": "^1.1.1",
+    "@vercel/kv": "^2.0.0",
     "classcat": "^5.0.4",
     "date-fns": "^3.2.0",
+    "date-fns-tz": "^3.1.3",
     "framer-motion": "^11.0.0",
     "next": "^14.0.0",
     "react": "^18.2.0",
@@ -30,11 +33,11 @@
     "@babel/eslint-parser": "^7.11.0",
     "@types/node": "^20.8.9",
     "@types/react": "^18.2.67",
-    "@typescript-eslint/eslint-plugin": "^7.0.0",
-    "@typescript-eslint/parser": "^7.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.0.0",
+    "@typescript-eslint/parser": "^8.0.0",
     "autoprefixer": "^10.4.16",
     "eslint": "^9.0.0",
-    "eslint-config-next": "14.2.3",
+    "eslint-config-next": "14.2.11",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.1",
     "postcss": "^8.4.31",
@@ -42,7 +45,7 @@
     "typescript": "^5.2.2"
   },
   "volta": {
-    "node": "18.20.3",
-    "yarn": "4.2.2"
+    "node": "20.17.0",
+    "yarn": "4.4.1"
   }
 }
