@@ -1,53 +1,68 @@
 import type { FC } from 'react'
+import { formatDate } from '@lib/formatDate'
+import { calculateDuration } from '@lib/calculateDuration'
 
-const experiences: {
+type Experience = {
   title: string
-  terms: string
+  startDate: string
+  endDate: string | null
   jobTitle: string
-}[] = [
+}
+
+const experiences: Experience[] = [
   {
     title: 'Crafts inc',
-    terms: '2017.1 ~ 2017.7',
+    startDate: '2017-01',
+    endDate: '2017-07',
     jobTitle: 'Interpreter / Translator',
   },
   {
     title: 'ALIVE inc',
-    terms: '2017.12 ~ 2019.3',
+    startDate: '2017-12',
+    endDate: '2019-03',
     jobTitle: 'Web Creator',
   },
   {
     title: 'Maple Systems',
-    terms: '2019.4 ~ 2019.12',
+    startDate: '2019-04',
+    endDate: '2019-12',
     jobTitle: 'System Engineer',
   },
   {
     title: 'Asia Quest',
-    terms: '2020.1 ~ 2021.10',
+    startDate: '2020-01',
+    endDate: '2021-10',
     jobTitle: 'Backend Developer',
   },
   {
     title: '■ ■ ■ ■ ■',
-    terms: '2021.11 ~ present',
-    jobTitle: 'Frontend Developer / Backend Developer'
+    startDate: '2021-11',
+    endDate: null,
+    jobTitle: 'Product Developer (Frontend Developer / Backend Developer)',
   },
 ]
 
 export const Step: FC = () => {
   return (
     <ul className={'w-[90%] mx-auto flex flex-col gap-y-[20px]'}>
-      {
-        experiences.map(ex => {
-          return (
-            <li key={ex.title}>
-              <dl className={'flex flex-col gap-y-[10px]'}>
-                <dt>{ex.title}</dt>
-                <dd className={'ml-[30px]'}>{ex.jobTitle}</dd>
-                <dd className={'ml-[30px]'}>{ex.terms}</dd>
-              </dl>
-            </li>
-          )
-        })
-      }
+      {experiences.map((ex) => {
+        const formattedStartDate = formatDate(ex.startDate)
+        const formattedEndDate = ex.endDate ? formatDate(ex.endDate) : 'present'
+        const duration = calculateDuration(ex.startDate, ex.endDate)
+
+        return (
+          <li key={ex.title}>
+            <div className={'flex flex-col gap-y-[10px]'}>
+              <div>{ex.title}</div>
+              <div className={'ml-[30px]'}>{ex.jobTitle}</div>
+              <div className={'ml-[30px]'}>
+                {`${formattedStartDate} ~ ${formattedEndDate}`}
+                <span className="ml-2 text-sm text-gray-600">{duration}</span>
+              </div>
+            </div>
+          </li>
+        )
+      })}
     </ul>
   )
 }
